feat(template)!: add unittests, get rid of old ingress spec
The following patch is a breaking change. It remove the support of the legacy
ingress specification.

Furthermore, it extends the ingress as well as the service specification to make
it more confortable for administrators to customize the deployment for their
needs. For example to add additional annotations of the ingress as well as to
customize the port of the service.

Additionally a basic set of unittests has been implemented to avoid regression.

Further steps would be to extend the unittests for the other resources.

BREAKING CHANGE